fix(dev-env): aux.* files renamed to match windows constraints on reserved filenames
