+++ packages/core/src/types/model-data.ts
@@ -1,7 +1,7 @@
 import type { IsAny } from 'type-fest'
+import type * as aux from './_aux'
+import type { Any } from './_aux'
 import type { NonEmptyArray } from './_common'
-import type * as aux from './aux'
-import type { Any } from './aux'
 import type { _stage } from './const'
 import type { ModelGlobals } from './global'
 import type { DeploymentElement, DeploymentRelationship } from './model-deployment'
