+++ packages/core/src/types/model-deployment.ts
@@ -1,7 +1,7 @@
 import { isTruthy } from 'remeda'
+import type * as aux from './_aux'
+import type { AnyAux, Unknown } from './_aux'
 import type { Link } from './_common'
-import type * as aux from './aux'
-import type { AnyAux, Unknown } from './aux'
 import type { FqnRef } from './fqnRef'
 import type { AbstractRelationship, ElementStyle } from './model-logical'
 import type { Icon } from './scalar'
