+++ packages/core/src/types/model-dump.test-d.ts
@@ -1,6 +1,6 @@
 import { describe, expectTypeOf, it } from 'vitest'
-import type * as aux from './aux'
-import type { Aux, SpecAux } from './aux'
+import type * as aux from './_aux'
+import type { Aux, SpecAux } from './_aux'
 import { _stage } from './const'
 import type { AuxFromDump, LikeC4ModelDump, SpecificationDump, SpecTypesFromDump } from './model-dump'
