+++ packages/core/src/types/view-parsed.dynamic.ts
@@ -1,8 +1,8 @@
 import { isArray } from 'remeda'
 import type { MergeExclusive, Simplify } from 'type-fest'
+import type * as aux from './_aux'
+import type { AnyAux } from './_aux'
 import type { ExclusiveUnion } from './_common'
-import type * as aux from './aux'
-import type { AnyAux } from './aux'
 import type { _type } from './const'
 import type { ModelFqnExpr } from './expression-model'
 import { isStepEdgeId } from './scalar'
